Enhancement: don't corrupt standalone input files
The original sign4j.c application recognized a usage scenario where
there were separate -in and -out files on the signing command line. In
that scenario, it would not create a temporary file; instead, it would
modify the ZIP comment size in the input file and sign it to create the
output file. Unfortunately, this means the input file would be altered
(and left in a corrupt state) as part of the operation.

A person who specified separate -in and -out files is likely to expect
that the input file will remain unmodified. This change alters the
behavior to meet that expectation.